Added POST support; fixed minor issues
POST and GET now both work. I added an extensive comment
that includes an example call using curl. I fixed an issue
where the Set<String> used to find name collisions in
the files to be added to the zip archive was not reset
between calls. Also, the servlet now adds ".nc" when
netCDF3 files are requested, no matter what the type of
the source files.