Modified how operation is provided
The operation to perform was derived from the URL, using the
last part of the path. I dropped that in favor of using a query
string parameter so that this servlet will work with our json
code (at least in theory; I'm not sure how it'll work in practice).
I also fixed an issue where POST would redirect when the URL
did not end in a slash - this is good behavior for GET, but it
breaks POST because clients like curl don't expect POST to
get a redirect.