Build changes for the aggregation servlet
Fir my aggregation servlet JUnit tests, I used apache commons io and
so I added that jar to the lib dir and then to the build.xml file. The
commons io jar is only installed for the opendap.war file; there are
other targets in the build.xml file where it does not seem to be a
factor.