URI-encoded the href output by contents.xsl
This was part of the fix for the NASA/Raytheon issue where directories
that contain colons were not working. Part of the fix was to change how
URL/URIs are scrubbed. However, to get the HTML pages to work in a
browser, the href elements built by centents.xsl had to use 
encode-for-uri(@name) and not just @name.