Fixed the hyraxTest tests so they no longer depend on a specific server version number
I adopted the same fix I use in the BES for date/time stamps: use sed to look for
a specific regex and replace it with something that won't vary. In this case I
look for 'OPeNDAP Hyrax (.*)' and replace it with 'OPeNDAP Hyrax (Not.A.Release).'
This means the exsiting baselines will work (I added the fix in for building new
baselines as well). I did not look for the OLFS version string.