Force current URL protocol on download URLs
This addresses a [mixed-content policy](https://blog.chromium.org/2020/02/protecting-users-from-insecure.html) issue when the site is hosted via HTTPS but OPeNDAP is hosted under an insecure (http) reverse proxy.  When deployed in this scenario, OPeNDAP generates insecure URLs when it should be generating https.  Chrome recently started to block these which is why we just started to notice it.  

>Mixed Content: The site at 'https://example.com/' was loaded over a secure connection, but the file at >'https://example.com/opendap/Florence/Covered%2520Data%2520Snapshots/5/NDBC%2520Standard%2520Meteorological%2520data/32012h2018.nc.nc?>time%5B0:1:2353%5D,latitude%5B0:1:0%5D' was redirected through an insecure connection. This file should be served over HTTPS. This download has been blocked. See >https://blog.chromium.org/2020/02/protecting-users-from-insecure.html for more details.

This solution is to simply force the current URL protocol on the download links.