Besapi parameters mod (#167)
* Trying design ideas

* First pass at implmenting TransmitCoordinator

* This commit contains a major change to the BesApi in which
we are now passing a new parameter down the chain. This parameter
is an implementation of the TransmitCoordinator interface.
This allows the BES class to determine if a request retry might be
attempted by allowing the BES class to check and see if the
response stream is committed (i.e. bytes have already by transmitted
to the requesting client)

* Dropped old version of besTransaction (was commented out)